requirements.txt
absl-py
actionlib
angles
astor
bondpy
camera-calibration-parsers
catkin
controller-manager
controller-manager-msgs
cv-bridge
cycler
Cython
diagnostic-analysis
diagnostic-common-diagnostics
diagnostic-updater
dynamic-reconfigure
gast
gencpp
geneus
genlisp
genmsg
gennodejs
genpy
grpcio
h5py
interactive-markers
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver
laser-geometry
Markdown
matplotlib
message-filters
mock
mrcnn
numpy
opencv-python
pbr
pluginlib
protobuf
pycocotools
pyparsing
python-dateutil
python-qt-binding
qt-dotgraph
qt-gui
qt-gui-cpp
qt-gui-py-common
resource-retriever
rosbag
rosboost-cfg
rosclean
roscreate
rosgraph
roslaunch
roslib
roslint
roslz4
rosmake
rosmaster
rosmsg
rosnode
rosparam
rospy
rosservice
rostest
rostopic
rosunit
roswtf
rqt-action
rqt-bag
rqt-bag-plugins
rqt-console
rqt-dep
rqt-graph
rqt-gui
rqt-gui-py
rqt-image-view
rqt-joint-trajectory-controller
rqt-launch
rqt-logger-level
rqt-moveit
rqt-msg
rqt-nav-view
rqt-plot
rqt-pose-view
rqt-publisher
rqt-py-common
rqt-py-console
rqt-reconfigure
rqt-robot-dashboard
rqt-robot-monitor
rqt-robot-steering
rqt-runtime-monitor
rqt-rviz
rqt-service-caller
rqt-shell
rqt-srv
rqt-tf-tree
rqt-top
rqt-topic
rqt-web
rviz
scikit-learn
scipy
sensor-msgs
six
sklearn
smach
smach-ros
smclib
tensorboard==1.13.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor
tf
tf-conversions
tf2-geometry-msgs
tf2-kdl
tf2-py
tf2-ros
topic-tools
Werkzeug
wiimote
xacro
